requirements.txt
manim==0.18.1
Pillow==11.0.0
ply==3.11